pkg/config: isolate config subpackage
