+++ pkg/config/util.go
@@ -1,0 +1,50 @@
+// Copyright 2016 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package config
+
+import (
+	"time"
+
+	"github.com/pingcap/pd/pkg/typeutil"
+)
+
+func adjustString(v *string, defValue string) {
+	if len(*v) == 0 {
+		*v = defValue
+	}
+}
+
+func adjustUint64(v *uint64, defValue uint64) {
+	if *v == 0 {
+		*v = defValue
+	}
+}
+
+func adjustInt64(v *int64, defValue int64) {
+	if *v == 0 {
+		*v = defValue
+	}
+}
+
+func adjustFloat64(v *float64, defValue float64) {
+	if *v == 0 {
+		*v = defValue
+	}
+}
+
+func adjustDuration(v *typeutil.Duration, defValue time.Duration) {
+	if v.Duration == 0 {
+		v.Duration = defValue
+	}
+}
