+++ server/api/redirector.go
@@ -20,6 +20,7 @@ import (
 	"strings"
 
 	"github.com/ngaut/log"
+	"github.com/pingcap/pd/pkg/config"
 	"github.com/pingcap/pd/server"
 )
 
@@ -61,7 +62,7 @@ func (h *redirector) ServeHTTP(w http.ResponseWriter, r *http.Request, next http
 		return
 	}
 
-	urls, err := server.ParseUrls(strings.Join(leader.GetClientUrls(), ","))
+	urls, err := config.ParseUrls(strings.Join(leader.GetClientUrls(), ","))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
